Move Actions into new source file.
