+++ model/shared/src/main/scala/explore/undo/KIListMod.scala
@@ -8,7 +8,7 @@ import monocle.Getter
 import monocle.Iso
 import monocle.Lens
 
-class KIListMod[A, K](protected val keyLens: Lens[A, K])
+case class KIListMod[A, K](protected val keyLens: Lens[A, K])
     extends IndexedCollMod[KeyedIndexedList, Int, A, cats.Id, K] {
 
   override protected val valueLens: Lens[A, A] = Iso.id.asLens
